Add Broadway instrumentation
Add instrumentation based on Telemetry events. Broadway translates
itself[1] to a topology of GenStage processes, and that's what exposed
through its telemetry events[2].

The strategy taken was then to use each GenStage start/stop event to
create a NewRelic transaction boundary. This does create a bit of
indirection as Broadway callback module talks in terms of
`handle_message/3` and `handle_batch/4`, but on the good side we get
closer to what "is really happening", get a better mapping of telemetry
events to transactions. Also, the docs are extensive describing those
stages.

To aleviate the case of Processors - the GenStage works with batches,
but Broadway callback call each message individually - the
instrumentaton here creates separate segments/spans for each processed
message.

The Batcher processe was initially added here, but removed before
publishing this commit because 1-) it's usually quite fast as its logic
is simple 2-) there is not much the user can do about it, as there is
way to add custom code yet.

[1]: https://hexdocs.pm/broadway/0.6.0/architecture.html
[2]: https://hexdocs.pm/broadway/0.6.0/Broadway.html#module-telemetry