Add HTTPoison as optional dependency
We have a few libraries wrapping HTTPoison that optionally
instrument if NewRelic is present during compilation. Due the lack of
this optional dep here, the compilation process is not deterministic and
may end up with cases where both libraries are included (our wrapper and
NewRelic) and still have no instrumentation.